refactor: moved DesktopNav component to /component/navigation directory
Relates #67
